Renamed DataProviderError to InvalidConfigurationDataError
