+++ src/PrestaShopBundle/Form/Exception/InvalidConfigurationDataError.php
@@ -28,12 +28,58 @@ declare(strict_types=1);
 
 namespace PrestaShopBundle\Form\Exception;
 
-use PrestaShop\PrestaShop\Core\Data\AbstractTypedCollection;
-
-class DataProviderErrorCollection extends AbstractTypedCollection
+class InvalidConfigurationDataError
 {
-    protected function getType(): string
+    /**
+     * @var int
+     */
+    private $errorCode;
+
+    /**
+     * @var string
+     */
+    private $fieldName;
+
+    /**
+     * @var int|null
+     */
+    private $languageId;
+
+    /**
+     * InvalidConfigurationDataError constructor.
+     *
+     * @param int $errorCode
+     * @param string $fieldName
+     * @param int|null $languageId
+     */
+    public function __construct(int $errorCode, string $fieldName, ?int $languageId = null)
     {
-        return DataProviderError::class;
+        $this->errorCode = $errorCode;
+        $this->fieldName = $fieldName;
+        $this->languageId = $languageId;
+    }
+
+    /**
+     * @return int
+     */
+    public function getErrorCode(): int
+    {
+        return $this->errorCode;
+    }
+
+    /**
+     * @return string
+     */
+    public function getFieldName(): string
+    {
+        return $this->fieldName;
+    }
+
+    /**
+     * @return int|null
+     */
+    public function getLanguageId(): ?int
+    {
+        return $this->languageId;
     }
 }
